backend_blog/Blog_project/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-baton==2.1.3
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-jet==1.0.8
django-paypal==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
form==0.0.1
future==0.18.2
google-api-core==1.26.3
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
html5lib==1.1
httplib2==0.19.0
idna==2.10
mysqlclient==2.0.3
packaging==20.9
pdfkit==0.6.1
Pillow==8.1.2
protobuf==3.15.6
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2020.4
reportlab==3.5.67
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
wincertstore==0.2
xhtml2pdf==0.2.5
